+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## 0.18.1 - 2026-03-19
+### Added
+- add verification for price when no oracle
+
+### Updated
+- print PTLinear data
+
 ## 0.18.0 - 2026-03-19
 ### Updated
 - option to open PR requires price sources